fix runtime cython error on windows with current libs

+++ avaframe/com1DFA/DFAfunctionsCython.pyx
@@ -51,7 +51,6 @@ def computeForceC(cfg, particles, fields, dem, int frictType):
       force dictionary
   particles : dict
   """
-  # read input parameters
   cdef double enthRef = cfg.getfloat('enthRef')
   cdef double muSamosAt = cfg.getfloat('musamosat')
   cdef double tau0SamosAt = cfg.getfloat('tau0samosat')
@@ -121,7 +120,7 @@ def computeForceC(cfg, particles, fields, dem, int frictType):
   cdef double[:] uxArray = particles['ux']
   cdef double[:] uyArray = particles['uy']
   cdef double[:] uzArray = particles['uz']
-  cdef long[:] ID = particles['ID']
+  cdef long long[:] ID = particles['ID']
   cdef double[:] totalEnthalpyArray = particles['totalEnthalpy']
   cdef double[:, :] VX = fields['Vx']
   cdef double[:, :] VY = fields['Vy']
@@ -696,7 +695,7 @@ def updatePositionC(cfg, particles, dem, force, fields, int typeStop=0):
   cdef int restitutionCoefficient = dam['restitutionCoefficient']
   cdef int nIterDam = dam['nIterDam']
   cdef int nDamPoints = dam['nPoints']
-  cdef long[:] cellsCrossed = dam['cellsCrossed']
+  cdef long long[:] cellsCrossed = dam['cellsCrossed']
   cdef double[:] xFootArray = dam['x']
   cdef double[:] yFootArray = dam['y']
   cdef double[:] zFootArray = dam['z']
